// citpwawcl/src/app/ispezione/models/azione-ispezione.model.ts
export interface AzioneIspezione {
  cfUtenteAzione: string;
  denomUtenteAzione: string;
  descrizioneAzione: string;
  dtAzione: number;
  fkTipoAzione: number;
  idAccertamento: number;
  idAzione: number;
  idIspezione2018: number;
  idSanzione: number;
  idVerifica: number;
  nomeDoc: string;
  nomeDocOriginale: string;
  uidIndex: string;
}